refactor: - move parseTypedDateForDisplay() to DateTools.ts
